treat content_length as a signed off_t
we pass -1 if it's not present

+++ hyperfs-get.h
@@ -7,7 +7,7 @@ struct hyperfs_state;
 FILE *hyperget(
 	struct hyperfs_state *remote,
 	const char           *path,
-	size_t               *content_len);
+	off_t                *content_len);
 
 
 /* vim: set noexpandtab sts=0 sw=8 ts=8: */
